+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## Unreleased
 ### Added 
 - Support to view TraceLogs from within the extension
+- Added command to remove database name from SQL
 
 ### Changed
 - The extension now activates as soon as Azure Data Studio opens (earlier would trigger on commands or on SQL documents)